refactor(init): simplify model selector state
